+++ lib/crypto/tweetnacl.dart
@@ -37,6 +37,9 @@ class TweetNaCl {
   // Signature length
   static const int signatureLength = 64;
 
+  // Extended key's length
+  static const int extendedKeyLength = 64;
+  
   static final Uint8List _0 =
       Uint8List.fromList([0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]); //16
   static final Uint8List _9 = Uint8List.fromList([